chore: move pivot table component location
