+++ packages/frontend/src/components/common/PivotTable.tsx
